refactor: OutputView - view package로 이동
